requirements.txt
# Core Framework
django>=5.0,<6.0
djangorestframework>=3.14.0

# CORS headers for frontend communication
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# AI/LLM Integration
openai>=1.0.0  # For GPT models
anthropic>=0.25.0  # For Claude API
langchain>=0.1.0  # LLM orchestration
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
google-generativeai>=0.3.0  # Gemini API
langchain-google-genai>=0.0.5  # LangChain integration for Gemini

# Multi-Agent Framework
langgraph>=0.0.20  # For agent workflows and state management
crewai>=0.1.0  # Alternative multi-agent framework

# Web Scraping & Automation
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0  # Modern web automation

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0  # Excel support
python-docx>=1.1.0  # Word document support
PyPDF2>=3.0.0  # PDF processing
reportlab>=4.0.0  # PDF generation

# Email Integration
python-dotenv>=1.0.0  # Environment variables
emails>=0.6.0

# Task Queue & Background Jobs
celery>=5.3.0
redis>=5.0.0  # For Celery broker
django-celery-beat>=2.5.0  # Periodic tasks
django-celery-results>=2.5.0  # Task results backend

# API & Documentation
drf-spectacular>=0.27.0  # OpenAPI/Swagger documentation
pyyaml>=6.0  # YAML support for configs

# Authentication & Security
djangorestframework-simplejwt>=5.3.0  # JWT authentication
python-decouple>=3.8  # Settings management

# Monitoring & Logging
sentry-sdk>=1.40.0  # Error tracking (optional)

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0  # Test fixtures

# Code Quality
black>=23.12.0  # Code formatting
flake8>=7.0.0  # Linting
isort>=5.13.0  # Import sorting

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
pillow>=10.1.0  # Image processing